# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: JobStream API
  version: 1.0.0
  description: |
    API REST pour la recherche et gestion d'offres d'emploi.
    Contrats partagés entre Backend (Spring Boot) et Frontend (Angular).
    
    **Approche API-First**: Ce fichier est la source de vérité.
    Les DTOs et clients sont générés automatiquement à partir de cette spec.
servers:
  - url: http://localhost:8080
    description: Development server

# ============================================================================
# PATHS - Endpoints RESTful
# ============================================================================
paths:
  # --------------------------------------------------------------------------
  # Jobs - Recherche externe (Adzuna)
  # --------------------------------------------------------------------------
  /adzuna/jobs:
    get:
      summary: Rechercher des offres d'emploi
      operationId: searchJobs
      tags:
        - Adzuna
      parameters:
        - name: query
          in: query
          required: true
          description: Mot-clé de recherche
          schema:
            type: string
            minLength: 1
            maxLength: 200
            pattern: "^.*\\S.*$"
          example: "développeur java paris"
        - name: page
          in: query
          required: false
          description: Numéro de page
          schema:
            type: integer
            default: 1
            minimum: 1
            example: 1
        - name: limit
          in: query
          required: false
          description: Nombre de résultats par page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
            example: 20
        - name: location
          in: query
          required: false
          description: Localisation de recherche
          schema:
            type: string
            pattern: "^[a-zA-Z\\s\\-.,]+$"
            example: "paris"
      responses:
        '200':
          description: Liste des jobs trouvés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdzunaJobSearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /adzuna/jobs/{externalId}:
    get:
      summary: Détail d'un job externe
      operationId: getAdzunaJobById
      tags:
        - Adzuna
      parameters:
        - name: externalId
          in: path
          required: true
          description: ID externe du job (Adzuna)
          schema:
            type: string
          example: "job_123456"
      responses:
        '200':
          description: Détails du job
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDto'
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs/{id}:
    get:
      summary: Détail d'un job sauvegardé
      operationId: getJobById
      tags:
        - Job
      parameters:
        - name: id
          in: path
          required: true
          description: ID interne du job sauvegardé
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Détails du job sauvegardé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDto'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      summary: Supprimer un job sauvegardé
      operationId: deleteJob
      tags:
        - Job
      parameters:
        - name: id
          in: path
          required: true
          description: ID du job à supprimer
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Job supprimé
        '404':
          $ref: '#/components/responses/NotFound'
  /jobs:
    get:
      summary: Liste des jobs sauvegardés
      operationId: getJobs
      tags:
        - Job
      parameters:
        - name: pageable
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/Pageable'
      responses:
        '200':
          description: Liste paginée des jobs sauvegardés
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedJobResponse'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      summary: Sauvegarder un job
      operationId: saveJob
      tags:
        - Job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JobRequestDto'
      responses:
        '201':
          description: Job sauvegardé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobDto'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
# ============================================================================
# COMPONENTS - Schemas & Responses
# ============================================================================
components:
  schemas:
    JobRequestDto:
      type: object
      description: Job trouvé via l'API externe (à sauvegarder)
      required:
        - externalId
        - title
        - company
        - location
      properties:
        externalId:
          type: string
          description: ID externe (Adzuna)
          example: "job_123456"
        title:
          type: string
          example: "Développeur Java Senior"
        company:
          type: string
          example: "TechCorp France"
        location:
          type: string
          example: "Paris, France"
        description:
          type: string
          description: Description complète de l'offre
          example: "Nous cherchons un développeur Java expérimenté..."
        salaryMin:
          type: integer
          description: Salaire minimum
          example: 50000
        salaryMax:
          type: integer
          description: Salaire maximum
          example: 80000
        contractType:
          type: string
          description: Type de contrat
          example: "CDI"
        postedDate:
          type: string
          format: date
          description: Date de publication
          example: "2026-03-20"
        jobUrl:
          type: string
          description: URL originale de l'offre
          example: "https://example.com/job/123"
        requirements:
          type: array
          items:
            type: string
          description: Liste des exigences
          example: ["Java 17", "Spring Boot", "PostgreSQL"]
    JobDto:
      description: Job sauvegardé
      allOf:
        - $ref: '#/components/schemas/JobRequestDto'
        - type: object
          properties:
            id:
              type: integer
              format: int64
            savedAt:
              type: string
              format: date-time
              description: Date de sauvegarde
            updatedAt:
              type: string
              format: date-time
              description: Dernière mise à jour
    PagedJobResponse:
      type: object
      description: Réponse paginée pour les jobs sauvegardés
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/JobDto'
        page:
          type: integer
        size:
          type: integer
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
    Pageable:
      type: object
      properties:
        page:
          type: integer
          minimum: 0
        size:
          type: integer
          minimum: 1
        sort:
          type: array
          items:
            type: string

    # ----------------------------------------------------------------------
    # Response DTOs
    # ----------------------------------------------------------------------
    AdzunaJobSearchResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobDto'
        total:
          type: integer
          description: Nombre total de résultats
          example: 150
        count:
          type: integer
          description: Nombre de résultats retournés
          example: 15

    # ----------------------------------------------------------------------
    # Error DTOs
    # ----------------------------------------------------------------------
    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
          example: "Job not found"
        message:
          type: string
          example: "Job not found"
        date:
          type: string
          format: date-time

  # ----------------------------------------------------------------------
  # Responses
  # ----------------------------------------------------------------------
  responses:
    BadRequest:
      description: Requête invalide
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "BAD_REQUEST"
            message: "Invalid parameters"

    NotFound:
      description: Ressource non trouvée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "NOT_FOUND"
            message: "Job not found"

    Conflict:
      description: Ressource déjà existante
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "CONFLICT"
            message: "Job already saved"

    InternalError:
      description: Erreur serveur interne
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: "INTERNAL_ERROR"
            message: "An unexpected error occurred"